Update Seq.fold and add Seq.map (#2)
Seq.fold is updated to match Seq better (e.g. using `currentChildIndex`
instead of `lastChildIndex` to keep track of the current child), and
also returning the initial value of the accumulator when it has no
input.

Seq.map is also added as a special case of fold that only uses input
values and not the output of the preceding sibling, similarly to
Par.map.