Dur (#9)
Add a .dur() modifier to Par and Seq. This allows setting an absolute
duration for the container, either extending its duration (similar to
adding a Delay at the end) or cutting it off (meaning that a final Delay
may be shortened, or some other children may be pruned altogether). Dur
stacks with take in that take(n) still applies when a duration is set—so
that n children must succeed within the set duration.

Also update Delay so that it can only be greater than zero, replacing
illegal values with a default Instant(), _i.e._, ignoring them.

This involves a lot of code changes as well as test updates, and will
still require more changes to ensure that everything is consistent
(_e.g._, the semantics of zero-duration Seq/Par needs to be reviewed).

222 tests!