Instantiating the score (#11)
The score itself gets instantiated, as well as its children at the time
at which they are added. The addition time can be set with the extra
parameter `at`, otherwise the current deck time is used (or zero if no
deck is set yet). In order to keep track of the current time, both deck
and score point to the tape used for instantiation.

Async tests are added to test this feature, making use of two events
from the tape, `end` and `fail`, occurring when a child instance of the
score instance finishes with a value or with a failure.