Await and Effect (#13)
Introduce asynchronicity with Await and synchronous effects with Effect.
Effect is the same as Instant but is allowed to have side effects (which
has no further incidence currently, but should affect Par.take). Await
is also similar to Instant in that it wraps a function call, but this
time an async call. Its duration is therefore undefined until the call
returns; the duration is set according to the end time.

New tests, including a manual test, are added. More tests are required
as a follow-up to verify the behaviour of nested Par/Seq, including with
map and fold.