DOMEvent (#16)
Introduce DOMEvent as a specialised way to handle events from DOM
elements. This item is similar to Await except that it waits for an
event from a target to end, with the event object itself as its value.
The Deck provides support for this feature by centralising event
listeners, adding/removing as necessary, and translating event times to
logical times. The manual images test adds a button that needs to be
clicked for the images to show up.

Follow-up note: DOMEvent, like Await, should be repeatable, but repeat()
does not handle unresolved children correctly yet so this will all be
fixed in a forthcoming PR.