Error result (#22)
Instead of simply failing, items can have an error result. Five kinds of
errors are introduced: runtime errors (for Instant, Effect and Await;
these can occur if the corresponding function throws an exception),
input errors (for Par.map, Seq.fold and Seq.map; these can occur if the
input is not an array, or the items for the input cannot be created),
timeout errors (for Await and Event; there can occur when the item is
cut off before the call finishes or the event occurs), CancelError (for
any cancelled instance), and FailureError (for failible items).

When such an error occurs, the instance ends with its `error` property
set to the underlying error instead, instead of simply setting `failed`
to true.