1O01 Try (#33)
Error handling supported is added with Try(child, catch) which replicates the classic try/catch pattern. The child is first tried, and if it ends with a value, then Try also ends with a value. If it results in an error, the catch child is then tried with the error as its input and Try ends as that child ends. Try is not repeatable and its duration cannot be set.

Tests are added (reaching 300 tests), the README is updated, and the images exemple is modified to allow for failure when loading the images (with the timeout set to a ridiculously low value so that image loading often fails, but should sometimes succeed).