Patch (#35)
Introduce a patch object that is being edited by the patcher. It does
not do much yet but will be used to create timing items from boxes. This
is mostly an excuse to split the main patcher file into slightly more
separate kinds of objects—Box, Cord, Patcher, DrageEventListener, &c.