1O0C Assertion failures when Await fails asynchronously
Actually the problem arises when the async call fails after the Await has already ended (because of timeout), so the item ends twice. A couple of other tests that verify the behaviour of async failure as added as well, as well as a wrong URL for the images manual test.