1Q09 Text and Element boxes (#59)
Very rudimentary text and element boxes to display DOM nodes when
active. Text simply shows the text given as argument (anything after
“Text”), while Element creates an (empty) DOM element with the tag name
and attributes passed as argument (_e.g._, “Element img
src="https://placekitten.com/400/300"”). When running, the input for the
element is removed and replaced with the content when active.