1P0C Event box (#60)
Add an Event box that takes a *reference* to an element as its input and
an event name as its parameter. Connecting the outlet of an Element to
the inlet of an Event creates a reference cord, rendered as a dashed
instead of a solid line, which indicates that the element is not part of
the score—it stays on the page forever.

Also add a Window object that represents the window (so that its events
can be listened to).