1S0N Optionally include tape occurrences in dump (#89)
Add a flag to `dump()` so that tape occurrences corresponding to items
can also be shown. This is used for the dump in the patcher, and also
for tests like cancel/prune to verify that no stray occurrences are left
on the tape. Also useful for testing 1S0H (forthcoming).