1S0H Review effect of Set.cancel (#90)
Cancelling Set could fail if it was followed by another scheduled set
because the occurrence for the cancel was added after the occurrence for
the following set, undoing it. In order to ensure that occurrences are
added in the correct order when they happen at the same instant, we go
up the instance tree to verify which one should go first.