1R0B Make canvas scrollable (#105)
Compute the bounding rect of all the boxes in the canvas and set the
minimum width of the canvas element so that it shows scrollbars if it
gets larger than its container. Do this when boxes are added and move,
and not when removed (so that the box only grows). When boxes are moved
past the origin xwise or ywise, just shift everything to make our life
simpler (this can be improved later with real panning and zooming, but
we probably will want a real scene graph by that point).

Beware that now the canvas may have scrolled when using the drag event
listener.