1T0A Dyadic Score (#117)
Enable the second inlet to Score (and functions) when the first one is
connected, so that Score supports two inputs (both inputs are simply
added to the score). Mark it as variadic since a follow-up will allow to
add more inlets. Take care to re-add or not remove it when the box is
edited.