1W07 Customize event handler (#144)
Event has an optional child that gets instantiated in a zero-duration
interval and a separate tape which is run immediately. This allows e.g.
preventing the event (see the Drag example) and is similar to what will
happen with the ramp. The patcher adds an extra inlet to Event as well.