2409 Event duration within longer thread is wrong (#199)
Make the event listener aware of the duration of the event to know when
the end of the event occurs (rather than the end of the thread). Some
assertions were failing in the test file because of stale event
listeners, so the VM has a shutdown() method to stop listening to
events.