2102 Topological sort for correct queue priority (#201)
When an item is add()ed to the VM, sort all items in topological order
(also detecting cycles, which cannot be introduced yet anyway). The
order is then used as a tie-breaker when queuing threads at the same
time, making sure that a parent does not end before its children. Cutoff
and timeout threads have the same order as the thread that they are
interrupting with a 0.5 shift to allow the thread to end normally when
it reaches its maximum duration exactly.