240C Sub threads of Repeat do not repeat (#203)
We need to make sure that each iteration of Par has its own state,
including subthreads. This means that each iteration spawns a new copy
of every child thread, offset by the new begin time (this will be
mirrored in the timeline for repeated instances). Code for Par is
reviewed and should be slightly simpler as a result.